+++ STM32/Controller/movement.h
@@ -3,5 +3,5 @@
 #define MINIMUM_PERCENT_ACTIVE 0.05
 
 void standardize_directions(float* x_value, float* y_value);
-float standardize_y(float y_value);
-float standardize_x(float x_value);
+int standardize_y(float y_value);
+int standardize_x(float x_value);
